First Commit
The version is a bit rough, but it does the trick.  It adds a button
to the main toolbar in Outlook Desktop.  Pushing the button will
generate a link suitable for Todoist and that can be passed through to
an automation of some sort to add it to the task management system.